LayoutTests/imported/w3c/web-platform-tests/largest-contentful-paint/lcp-buffer-size.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>LargestContentfulPaint: buffer size is 150.</title>
<link rel="help" href="https://w3c.github.io/timing-entrytypes-registry/#registry">
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<section id="container"></section>
<script>
  // The timing entry types registry specifies a maxBufferSize of 150
  // for largest-contentful-paint entries, but generating 150 entries makes the test too slow,
  // so check for 20.
  const maxBufferSize = 150;
  const entriesToGenerate = 20;

  promise_test(async (t) => {
    assert_implements(window.LargestContentfulPaint, "LargestContentfulPaint is not implemented");

    const container = document.getElementById('container');

    // Generate progressively larger text elements to produce LCP candidates.
    // Each must be larger than the previous to trigger a new LCP entry.
    // Text elements are reliable LCP candidates (unlike plain background-color divs).
    for (let i = 0; i < entriesToGenerate; ++i) {
      const size = 10 + i;
      const p = document.createElement('p');
      p.style.cssText = `font-size: ${size}px; position: absolute; top: 0; left: 0;`;
      p.textContent = 'LCP candidate text';
      container.appendChild(p);

      // Each element needs its own rendering opportunity to be a separate LCP candidate.
      await new Promise(resolve => requestAnimationFrame(() => requestAnimationFrame(resolve)));
    }

    const entries = await new Promise(resolve => {
      new PerformanceObserver((entryList, observer) => {
        resolve(entryList.getEntries());
        observer.disconnect();
      }).observe({ type: 'largest-contentful-paint', buffered: true });
    });

    assert_equals(entries.length, entriesToGenerate, `LCP buffer size is at least ${entriesToGenerate}`);
  }, `LCP buffer size should be greater than one`);
</script>
</body>
